fix: extract method into renamed utils file
