+++ src/electron/utils/fileUtils.ts
@@ -1,5 +1,7 @@
 import path from "path";
 import { readFileSync } from "fs";
+import { writeFile, mkdir } from "fs/promises";
+import { logger } from "./logger.js";
 
 export function readFiles(filePaths: string[]): FileData[] | Error {
     const collectedFileData: FileData[] = [];
@@ -31,4 +33,34 @@ export function readFiles(filePaths: string[]): FileData[] | Error {
     }
 
     return collectedFileData;
+}
+
+export async function saveFile(fullFilePath: string, data: ArrayBuffer) {
+    const directoryPath = path.dirname(fullFilePath);
+
+    logger.info(`Received data for saving to path <${fullFilePath}>.`);
+
+    try {
+        await mkdir(directoryPath, { recursive: true });
+
+        const arrayBuffer = data;
+        const buffer = Buffer.from(arrayBuffer);
+
+        await writeFile(fullFilePath, buffer);
+
+        logger.info(`Successfully saved file to <${fullFilePath}>.`);
+        return;
+    } catch (error) {
+        logger.error(
+            `Failed to save file <${fullFilePath}>! Details: <${
+                (error as Error).message
+            }>.`,
+            error,
+        );
+        return new Error(
+            `Failed to save file <${fullFilePath}>! Details: <${
+                (error as Error).message
+            }>.`,
+        );
+    }
 }
